ath10k: Add better support for CT firmware and 9887 NIC.
The 9887 NIC is a single-stream device, but firmware is compiled
for 3x3, so add a fixup when calculating the rateset.  In addition,
I cannot at this time get VHT rates to work on 9887 (HT works fine).
I don't know why that is...the firmware appears to be doing the right
thing, the packet just never hits the air when VHT rates are used.

Signed-off-by: Ben Greear <greearb@candelatech.com>